requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
chromadb==0.4.24
pymupdf==1.26.1
httpx==0.25.2 
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
numpy==1.24.3
tiktoken==0.5.2
schedule==1.2.0
python-dotenv==1.0.0
redis==5.0.1